Move HTTP output code from conky.cc to display-http
